requirements.txt
torch
torchvision
numpy
pandas
opencv-python
matplotlib
scikit-learn
streamlit
Pillow